refactor: extract schema conversion utilities into schema_utils.py
Move _simplify_schema(), _get_simplified_schema(), and
_output_fields_to_output_schema() from server.py into a dedicated
schema_utils.py module. These are pure utility functions with no
dependencies on server logic, improving separation of concerns and
making them independently importable/testable.

No behavioral change — all 137 existing tests pass.
